% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elizatemplate.R
\name{article}
\alias{article}
\title{Article}
\usage{
article(..., md_extensions = c("-autolink_bare_uris"))
}
\arguments{
\item{...}{Optional arguments passed to \code{\link{pdf_document}}}

\item{md_extensions}{Markdodown extensions, cf. \code{\link{pdf_document}}}
}
\description{
Formatting an article for self-archiving
}
\details{
The fonction is called by the Markdown model Basic article
}
