% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elizatemplate.R
\name{Tricoter}
\alias{Tricoter}
\alias{TricoterTout}
\alias{TricoterArticle}
\title{Tricoter}
\usage{
TricoterTout(destination = "docs")

TricoterArticle(destination = "docs")
}
\arguments{
\item{destination}{Destination file of the documents}
}
\description{
Creating all documents from the models
}
\details{
Used to verify that the models function well
}
