% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elizatemplate.R
\name{Tricoter}
\alias{Tricoter}
\alias{TricoterTout}
\alias{TricoterArticle}
\title{Tricoter}
\usage{
TricoterTout(destination = "docs")

TricoterArticle(destination = "docs")
}
\arguments{
\item{destination}{Dossier de destination des documents}
}
\description{
Création des tous les documents à partir des modèles
}
\details{
Utilisé pour vérifier le bon fonctionnement des modèles
}
